flink/glucose_stream_job.py
# PyFlink streaming job (skeleton)
from pyflink.datastream import StreamExecutionEnvironment

env = StreamExecutionEnvironment.get_execution_environment()

# Kafka source and sink would be configured here

env.execute("Glucose Streaming Job")
