requirements.txt
pandas
openpyxl
kafka-python
pyarrow
